TASK: Lock events table before committing
There are situations where two processes try to write data to the event
store simultaneously.

|sequence number|event                 |source        |
|1              |do first thing        |first process |
|2              |do second thing       |first process |
|3              |do something different|second process|
|4              |do third thing        |first process |
|5              |do fourth thing       |first process |

The second process starts second but will finish first.

So there is a point in time the event store looks like this:

|sequence number|event                 |source        |
|3              |do something different|second process|

If now a listener starts catching up, it will be remered as
highest applied sequence number = 3.

As soon as the first process finishes, another catch up happens starting
at sequence number 4.
So in this very situation, sequence numbers 1 and 2 are recorded and
even recorded in the right order but never played.